feat: Add `oneOption` extension method to MappedAsyncPagingIterable.
In previous releases we added the `nextOption` method which would return an
`Future[Option[(T, MappedAsyncPagingIterable)]]`. This was done to provide a similar
API to Pager, in which the client code could iterate over results while keeping
track of the mutated paging iterable, which would change when fetching the next page.

An example can be seen on the tests:

```scala
        result <- "SELECT * FROM hotels".toCQLAsync.prepareUnit
          .as[Hotel]
          .map(_.withPageSize(2))
          .executeAsync()

        Some((hotelA, iteratorA)) <- result.nextOption()
        Some((hotelB, iteratorB)) <- iteratorA.nextOption() // we use 'iteratorA', not 'result'
        Some((hotelC, iteratorC)) <- iteratorB.nextOption()
        Some((hotelD, iteratorD)) <- iteratorC.nextOption()
        Some((hotelE, iteratorE)) <- iteratorD.nextOption()
        lastResult <- iteratorE.nextOption()
```

This is cumbersome when we want _only_ the first result, for example when
we use a `LIMIT 1` in the query, which is where `oneOption` becomes useful:

```scala
        result <- "SELECT * FROM hotels LIMIT 1".toCQLAsync.prepareUnit
          .as[Hotel]
          .executeAsync()

        firstResult = result.oneOption  // this should be Some
        nextResult  = result.oneOption  // this will always be None
```

`oneOption` is no longer providing a `Future` but only an `Option` as the
page is already fetched.